Did a small performance test to see if the GPU is busy all the time

+++ src/awkward/_v2/_connect/cuda/cuda_kernels/cuda_common.cu
@@ -28,4 +28,4 @@ typedef unsigned int       uint_least32_t;
 typedef unsigned long long uint_least64_t;
 typedef unsigned long long uintmax_t;
 
-#define RAISE_ERROR(ERROR_KERNEL_CODE) atomicMin(err_code, invocation_index * (1 << ERROR_BITS) + (ERROR_KERNEL_CODE));
+#define RAISE_ERROR(ERROR_KERNEL_CODE) atomicMin(err_code, invocation_index * (1 << ERROR_BITS) + (int)(ERROR_KERNEL_CODE));
